chore: Move Popover component to UI directory
