--- packages/webapp-libs/webapp-core/src/components/popover/index.ts
@@ -1,3 +1,0 @@
-export * from './popover.component';
-export * from './popoverContent';
-export * from './popoverTrigger';
