--- packages/webapp-libs/webapp-core/src/components/popover/popover.component.tsx
@@ -1,5 +1,0 @@
-import * as PopoverPrimitive from '@radix-ui/react-popover';
-
-const Popover = PopoverPrimitive.Root;
-
-export { Popover };
