--- packages/webapp-libs/webapp-core/src/components/popover/popoverContent/index.ts
@@ -1,1 +1,0 @@
-export * from './popoverContent';
