--- packages/webapp-libs/webapp-core/src/components/popover/popoverTrigger/index.ts
@@ -1,1 +1,0 @@
-export * from './popoverTrigger.component';
