--- packages/webapp-libs/webapp-core/src/components/popover/popoverTrigger/popoverTrigger.component.tsx
@@ -1,5 +1,0 @@
-import * as PopoverPrimitive from '@radix-ui/react-popover';
-
-const PopoverTrigger = PopoverPrimitive.Trigger;
-
-export { PopoverTrigger };
